scripts/etl-pipeline/tenders/requirements.txt
# requirements.txt
python-dotenv==1.0.0
supabase==2.3.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.4
python-dateutil==2.8.2